+++ packages/components/atoms/f-button/test/component/f-button.component.desktop.spec.js
@@ -1,4 +1,3 @@
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const Button = require('../../test-utils/component-objects/f-button.component');
 
 let button;
@@ -9,11 +8,9 @@ describe('f-button component tests', () => {
         button = new Button();
         button.withQuery('knob-Button Type', 'primary');
         button.withQuery('knob-Button Size', 'medium');
-        const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
 
         // Act
-        button.open(pageUrl);
-        button.waitForActionComponent();
+        button.load();
 
         // Assert
         expect(button.isActionComponentDisplayed()).toBe(true);
@@ -26,11 +23,8 @@ describe('f-button component tests', () => {
             .withQuery('knob-href', 'link')
             .withQuery('knob-Button Size', 'medium');
 
-        const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
         // Act
-        button.open(pageUrl);
-        button.waitForLinkComponent();
+        button.load('link');
 
         // Assert
         expect(button.isLinkComponentDisplayed()).toBe(true);