+++ packages/components/atoms/f-error-message/test-utils/component-objects/f-error-message.component.js
@@ -5,6 +5,10 @@ module.exports = class ErrorMessage extends Page {
         super('atom', 'error-message-component');
     }
 
+    load () {
+        super.load(this.component);
+    }
+
     get component () { return $('[data-test-id="error-message-component"]'); }
 
     open (url) {