+++ packages/components/atoms/f-popover/test-utils/component-objects/f-popover.component.js
@@ -8,6 +8,10 @@ module.exports = class Popover extends Page {
 
     get component () { return $(COMPONENT); }
 
+    load () {
+        super.load(this.component);
+    }
+
     open (url) {
         super.open(url);
     }