+++ packages/components/molecules/f-searchbox/test/accessibility/axe-accessibility.spec.js
@@ -1,4 +1,3 @@
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const { getAccessibilityTestResults } = require('../../../../../../test/utils/axe-helper');
 
 const Searchbox = require('../../test-utils/component-objects/f-searchbox.component');
@@ -7,9 +6,7 @@ const searchbox = new Searchbox();
 
 describe('Accessibility tests', () => {
     beforeEach(() => {
-        const pageUrl = buildUrl(searchbox.componentType, searchbox.componentName, searchbox.path);
-        searchbox.open(pageUrl);
-        searchbox.waitForComponent();
+        searchbox.load();
     });
     it('a11y - should test f-searchbox component WCAG compliance', () => {
         // Act