+++ packages/components/molecules/f-skeleton-loader/test-utils/component-objects/f-skeletonLoader.component.js
@@ -8,6 +8,10 @@ module.exports = class SkeletonLoader extends Page {
 
     get component () { return $(COMPONENT); }
 
+    load () {
+        super.load(this.component);
+    }
+
     open (url) {
         super.open(url);
     }