+++ packages/components/organisms/f-checkout/src/components/AgeVerification.vue
@@ -1,6 +1,5 @@
 <template>
     <card
-        is-rounded
         is-page-content-wrapper
         card-heading-position="center"
         has-outline
@@ -116,7 +115,7 @@ export default {
 
         months () {
             return new Array(12).fill(0).map((_, i) => {
-                const date = new Date(`${i + 1}/1`);
+                const date = new Date(new Date().getFullYear(), i);
 
                 return {
                     text: date.toLocaleDateString(undefined, { month: 'long' }),