+++ packages/components/organisms/f-checkout/src/components/_tests/Checkout.test.js
@@ -13,7 +13,8 @@ import {
     ERROR_CODE_FULFILMENT_TIME_INVALID,
     ERROR_CODE_FULFILMENT_TIME_UNAVAILABLE,
     TENANT_MAP,
-    CHECKOUT_ERROR_FORM_TYPE
+    CHECKOUT_ERROR_FORM_TYPE,
+    HEADER_LOW_VALUE_ORDER_EXPERIMENT
 } from '../../constants';
 import VueCheckout from '../Checkout.vue';
 import EventNames from '../../event-names';
@@ -43,6 +44,14 @@ const localVue = createLocalVue();
 
 localVue.use(VueI18n);
 localVue.use(Vuex);
+
+jest.mock('../../services/analytics', () => jest.fn().mockImplementation(() => ({
+    trackFormInteraction: jest.fn(),
+    trackInitialLoad: jest.fn(),
+    trackFormErrors: jest.fn(),
+    trackDialogEvent: jest.fn(),
+    trackLowValueOrderExperiment: jest.fn()
+})));
 
 const $v = {
     customer: {
@@ -1212,13 +1221,11 @@ describe('Checkout', () => {
     describe('mounted ::', () => {
         let initialiseSpy;
         let setAuthTokenSpy;
-        let trackInitialLoadSpy;
         let wrapper;
 
         beforeEach(() => {
             initialiseSpy = jest.spyOn(VueCheckout.methods, 'initialise');
             setAuthTokenSpy = jest.spyOn(VueCheckout.methods, 'setAuthToken');
-            trackInitialLoadSpy = jest.spyOn(VueCheckout.methods, 'trackInitialLoad');
 
             wrapper = shallowMount(VueCheckout, {
                 store: createStore(),
@@ -1247,7 +1254,7 @@ describe('Checkout', () => {
             await wrapper.vm.initialise();
 
             // Assert
-            expect(trackInitialLoadSpy).toHaveBeenCalled();
+            expect(wrapper.vm.checkoutAnalyticsService.trackInitialLoad).toHaveBeenCalled();
         });
 
         it('should emit `CheckoutMounted` event', async () => {
@@ -1976,14 +1983,11 @@ describe('Checkout', () => {
                 });
 
                 it('should make a call to `trackFormErrors`', () => {
-                    // Arrange
-                    const trackFormErrorsSpy = jest.spyOn(wrapper.vm, 'trackFormErrors');
-
                     // Act
                     wrapper.vm.handleNonFulfillableCheckout();
 
                     // Assert
-                    expect(trackFormErrorsSpy).toHaveBeenCalled();
+                    expect(wrapper.vm.checkoutAnalyticsService.trackFormErrors).toHaveBeenCalled();
                 });
 
                 it('should emit `CheckoutUpdateFailure` event', async () => {
@@ -2079,6 +2083,36 @@ describe('Checkout', () => {
 
                         // Assert
                         expect(reloadAvailableFulfilmentTimesIfOutdatedSpy).toHaveBeenCalled();
+                    });
+                });
+
+                describe('when `updateCheckout` call returns headers', () => {
+                    // Arrange
+                    const checkoutResponseHeaders = {
+                        [HEADER_LOW_VALUE_ORDER_EXPERIMENT]: 'test'
+                    };
+
+                    afterEach(() => {
+                        jest.clearAllMocks();
+                    });
+
+                    it('should call `trackLowValueOrderExperiment` with `checkoutResponseHeaders`', async () => {
+                        // Arrange
+                        wrapper = mount(VueCheckout, {
+                            store: createStore(defaultCheckoutState, {
+                                ...defaultCheckoutActions,
+                                updateCheckout: jest.fn(() => Promise.resolve(checkoutResponseHeaders))
+                            }),
+                            i18n,
+                            localVue,
+                            propsData
+                        });
+
+                        // Act
+                        await wrapper.vm.handleUpdateCheckout();
+
+                        // Assert
+                        expect(wrapper.vm.checkoutAnalyticsService.trackLowValueOrderExperiment).toHaveBeenCalledWith(checkoutResponseHeaders);
                     });
                 });
             });
@@ -2310,6 +2344,85 @@ describe('Checkout', () => {
                         result.rejects.toThrow(errorMessage);
                     });
                 });
+            });
+        });
+
+        describe('mapCheckoutUpdateRequest', () => {
+            let wrapper;
+            let getMappedDataForUpdateCheckoutSpy;
+
+            beforeEach(() => {
+                getMappedDataForUpdateCheckoutSpy = jest.spyOn(VueCheckout.methods, 'getMappedDataForUpdateCheckout');
+
+                wrapper = shallowMount(VueCheckout, {
+                    store: createStore({
+                        ...defaultCheckoutState,
+                        customer: {
+                            ...defaultCheckoutState.customer,
+                            dateOfBirth: 'Thu Jul 05 1990 00:00:00 GMT+0100 (British Summer Time)'
+                        }
+                    }),
+                    i18n,
+                    localVue,
+                    propsData,
+                    mocks: {
+                        $logger,
+                        $cookies
+                    }
+                });
+            });
+
+            afterEach(() => {
+                jest.clearAllMocks();
+            });
+
+            it('should map the request successfully', async () => {
+                // Act
+                const mappedRequest = await wrapper.vm.getMappedDataForUpdateCheckout();
+
+                // Assert
+                expect(getMappedDataForUpdateCheckoutSpy).toHaveBeenCalled();
+                expect(mappedRequest[0].value).toEqual({
+                    dateOfBirth: 'Thu Jul 05 1990 00:00:00 GMT+0100 (British Summer Time)',
+                    firstName: 'John',
+                    lastName: 'Smith',
+                    phoneNumber: '0711111111'
+                });
+
+                expect(mappedRequest[1].value).toEqual({
+                    location: {
+                        address: {
+                            administrativeArea: undefined,
+                            lines: [
+                                '1 Bristol Road',
+                                'Flat 1'
+                            ],
+                            locality: 'Bristol',
+                            postalCode: 'BS1 1AA'
+                        },
+                        geolocation: null
+                    },
+                    table: {},
+                    time: {
+                        asap: undefined,
+                        scheduled: {
+                            from: '',
+                            to: ''
+                        }
+                    }
+                });
+            });
+
+            it('should map the request for age verification only successfully', async () => {
+                // Act
+                const mappedRequest = await wrapper.vm.getMappedDataForUpdateCheckout({ ageVerificationOnly: true });
+
+                // Assert
+                expect(getMappedDataForUpdateCheckoutSpy).toHaveBeenCalled();
+                expect(mappedRequest[0].value).toEqual({
+                    dateOfBirth: 'Thu Jul 05 1990 00:00:00 GMT+0100 (British Summer Time)'
+                });
+                expect(mappedRequest[1].value).toBeNull();
             });
         });
 
@@ -2662,6 +2775,26 @@ describe('Checkout', () => {
                     // Assert
                     expect(getGeoLocationSpy).toHaveBeenCalledWith(expected);
                 });
+
+                it('should not call `getGeoLocation` if the service type is not delivery', async () => {
+                    // Arrange
+                    wrapper = mount(VueCheckout, {
+                        store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                        i18n,
+                        localVue,
+                        propsData,
+                        mocks: {
+                            $logger,
+                            $cookies
+                        }
+                    });
+
+                    // Act
+                    await wrapper.vm.lookupGeoLocation();
+
+                    // Assert
+                    expect(getGeoLocationSpy).toHaveBeenCalledTimes(0);
+                });
             });
 
             describe('when `getGeoLocation` request fails', () => {
@@ -2911,7 +3044,6 @@ describe('Checkout', () => {
             let error;
             let eventToEmit;
             let logInvokerSpy;
-            let trackFormInteractionSpy;
             let scrollToElementSpy;
 
             beforeEach(() => {
@@ -2933,7 +3065,6 @@ describe('Checkout', () => {
                 });
 
                 logInvokerSpy = jest.spyOn(wrapper.vm, 'logInvoker');
-                trackFormInteractionSpy = jest.spyOn(wrapper.vm, 'trackFormInteraction');
                 scrollToElementSpy = jest.spyOn(wrapper.vm, 'scrollToElement');
             });
 
@@ -2987,7 +3118,7 @@ describe('Checkout', () => {
                 wrapper.vm.handleErrorState(error);
 
                 // Assert
-                expect(trackFormInteractionSpy).toHaveBeenCalledWith({ action: 'error', error: `error_${error.errorCode}-${error.message}` });
+                expect(wrapper.vm.checkoutAnalyticsService.trackFormInteraction).toHaveBeenCalledWith({ action: 'error', error: `error_${error.errorCode}-${error.message}` });
             });
 
             it('should call `trackFormInteraction` without an `errorCode` when it does not exist', () => {
@@ -2998,7 +3129,7 @@ describe('Checkout', () => {
                 wrapper.vm.handleErrorState(error);
 
                 // Assert
-                expect(trackFormInteractionSpy).toHaveBeenCalledWith({ action: 'error', error: `error_${error.message}` });
+                expect(wrapper.vm.checkoutAnalyticsService.trackFormInteraction).toHaveBeenCalledWith({ action: 'error', error: `error_${error.message}` });
             });
 
             it('should call `scrollToElement` with the `errorMessage` element', async () => {
@@ -3267,7 +3398,7 @@ describe('Checkout', () => {
                 it('should make a call to `trackFormInteraction` so we can track the action type `submit`', async () => {
                     // Arrange
                     isFormValidSpy.mockReturnValue(true);
-                    const trackFormInteractionSpy = jest.spyOn(VueCheckout.methods, 'trackFormInteraction');
+
                     const wrapper = mount(VueCheckout, {
                         store: createStore(),
                         i18n,
@@ -3284,7 +3415,7 @@ describe('Checkout', () => {
                     await wrapper.vm.onFormSubmit();
 
                     // Assert
-                    expect(trackFormInteractionSpy).toHaveBeenCalledWith({
+                    expect(wrapper.vm.checkoutAnalyticsService.trackFormInteraction).toHaveBeenCalledWith({
                         action: 'submit'
                     });
                 });
@@ -3417,7 +3548,6 @@ describe('Checkout', () => {
             let isFormValidSpy;
             let mockValidationState;
             let getFormValidationStateSpy;
-            let trackFormInteractionSpy;
 
             beforeEach(() => {
                 mockValidationState = {
@@ -3433,7 +3563,6 @@ describe('Checkout', () => {
                 getFormValidationStateSpy = jest.spyOn(validations, 'getFormValidationState');
                 getFormValidationStateSpy.mockReturnValue(mockValidationState);
                 isFormValidSpy = jest.spyOn(VueCheckout.methods, 'isFormValid');
-                trackFormInteractionSpy = jest.spyOn(VueCheckout.methods, 'trackFormInteraction');
             });
 
             it('should exist', () => {
@@ -3503,7 +3632,7 @@ describe('Checkout', () => {
                     await wrapper.vm.onFormSubmit();
 
                     // Assert
-                    expect(trackFormInteractionSpy).toHaveBeenCalledWith({
+                    expect(wrapper.vm.checkoutAnalyticsService.trackFormInteraction).toHaveBeenCalledWith({
                         action: 'inline_error',
                         error: mockValidationState.invalidFields.toString()
                     });
@@ -3529,7 +3658,7 @@ describe('Checkout', () => {
                     await wrapper.vm.onFormSubmit();
 
                     // Assert
-                    expect(trackFormInteractionSpy).toHaveBeenCalledWith({
+                    expect(wrapper.vm.checkoutAnalyticsService.trackFormInteraction).toHaveBeenCalledWith({
                         action: 'error',
                         error: ANALYTICS_ERROR_CODE_INVALID_MODEL_STATE
                     });
@@ -4096,6 +4225,29 @@ describe('Checkout', () => {
 
                 // Assert
                 expect(wrapper.vm.availableFulfilmentTimesKey).toEqual(1);
+            });
+        });
+
+        describe('handleDialogCreation ::', () => {
+            const event = {
+                code: 'DuplicateOrder',
+                isDuplicateOrderError: true
+            };
+
+            it('should call `trackDialogEvent` with passed event', () => {
+                // Arrange
+                const wrapper = mount(VueCheckout, {
+                    store: createStore(),
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
+                // Act
+                wrapper.vm.handleDialogCreation(event);
+
+                // Assert
+                expect(wrapper.vm.checkoutAnalyticsService.trackDialogEvent).toHaveBeenCalledWith(event);
             });
         });
     });