+++ packages/components/organisms/f-header/test-utils/component-objects/f-header.component.js
@@ -46,6 +46,10 @@ module.exports = class Header extends Page {
         this.countryValue = this.navigation.countrySelector.countries.filter(element => element.getAttribute('data-test-id').includes(country))[0];
     }
 
+    load () {
+        super.load(this.component);
+    }
+
     open (url) {
         super.open(url);
     }