+++ packages/services/f-analytics/src/services/analytics.service.js
@@ -1,16 +1,81 @@
+import analyticsModule from '../store/analytics.module';
+import defaultOptions from '../defaultOptions';
 import {
     mapPlatformData,
     mapUserData,
-    mapPageData
+    mapPageData,
+    mapServerSidePlatformData
 } from './analytics.mapper';
 
 const isDataLayerPresent = () => typeof (window) !== 'undefined' && window.dataLayer;
+
+const prepareServerSideValues = (store, req, options) => {
+    // Only available server side
+    if (typeof (window) === 'undefined') {
+        let platformData = { ...store.state[`${options.namespace}`].platformData };
+
+        platformData = mapServerSidePlatformData({ platformData, req });
+
+        store.dispatch(`${options.namespace}/updatePlatformData`, platformData);
+    }
+};
+
+const preparePageTags = options => {
+    // Only add tags if clientside and if not already added
+    if (typeof (window) !== 'undefined' && !window.dataLayer) {
+        const queryString = options.auth ? `&gtm_auth=${options.auth}&gtm_preview=${options.preview}&gtm_cookies_win=${options.cookiesWin}` : '';
+
+        // See : https://developers.google.com/tag-manager/quickstart
+        const headJsGtmTag = `(function (w, d, s, l, i) {
+                                    w[l] = w[l] || [];
+                                    w[l].push({
+                                        'gtm.start': new Date().getTime(),
+                                        event: 'gtm.js'
+                                    });
+                                    var f = d.getElementsByTagName(s)[0],
+                                        j = d.createElement(s),
+                                        dl = l != 'dataLayer' ? '&l=' + l : '',
+                                        qs = '${queryString}';
+                                    j.async = true;
+                                    j.src = 'https://www.googletagmanager.com/gtm.js?id=' + i + dl + qs;
+                                    f.parentNode.insertBefore(j, f);
+                                })(window, document, 'script', 'dataLayer', '${options.id}');`;
+        const script = document.createElement('script');
+        script.text = headJsGtmTag;
+        document.head.prepend(script);
+
+        // No JS option
+        const iframeTag = document.createElement('iframe');
+        iframeTag.src = `https://www.googletagmanager.com/ns.html?id=${options.id}`;
+        iframeTag.setAttribute('height', 0);
+        iframeTag.setAttribute('width', 0);
+        iframeTag.style.display = 'none';
+        iframeTag.style.visibility = 'hidden';
+        const noScript = document.createElement('noscript');
+        noScript.appendChild(iframeTag);
+        document.body.prepend(noScript);
+    }
+};
 
 export default class AnalyticService {
     constructor (store, req, options) {
         this.store = store;
         this.req = req;
-        this.options = options;
+        this.options = {
+            ...defaultOptions,
+            ...options
+        };
+
+        preparePageTags(this.options);
+        store.registerModule(this.options.namespace, analyticsModule, { preserveState: !!store.state[`${this.options.namespace}`] });
+        prepareServerSideValues(this.store, this.req, this.options);
+
+        // Flush any previously stored (serverside) events
+        this.pushEvent();
+    }
+
+    getOptions () {
+        return this.options;
     }
 
     setOptions ({ featureName = this.options.featureName, locale = this.options.locale } = {}) {
@@ -25,8 +90,7 @@ export default class AnalyticService {
         platformData = mapPlatformData({
             platformData,
             featureName: featureName || this.options.featureName,
-            locale: locale || this.options.locale,
-            req: this.req
+            locale: locale || this.options.locale
         });
 
         this.store.dispatch(`${this.options.namespace}/updatePlatformData`, platformData);
@@ -61,7 +125,6 @@ export default class AnalyticService {
     pushPageData ({
         pageName,
         conversationId,
-        requestId,
         httpStatusCode,
         customFields
     } = {}) {
@@ -69,10 +132,8 @@ export default class AnalyticService {
 
         pageData = mapPageData({
             pageData,
-            featureName: this.options.featureName,
             pageName,
             conversationId,
-            requestId,
             httpStatusCode,
             req: this.req
         });