+++ packages/components/organisms/f-takeawaypay-activation/CHANGELOG.md
@@ -3,6 +3,23 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v2.1.0
+------------------------------
+*October 05, 2021*
+
+### Added
+- Visual regression tests
+
+v2.0.0
+------------------------------
+*October 5, 2021*
+
+### Changed
+- Updated version of `f-button` and `f-card` to include icing phase 2 changes.
+
+### Removed
+- `is-rounded` prop from `f-card` component call as the card is now rounded by default.
+
 
 v1.2.0
 ------------------------------